[incomplete] Add optional topic update loop for minecraft
This is currently inop. The plugin structure doesn't provide plugins
with a way to send messages outside of receiving the bot context from
the message stream. It's possible that we could store a reference
in the plugin to the first instance of that context, but that's
pretty gnarly.

Any ideas? Does this seem reasonable outside of the `TODO`s and
the above issue?