Add mention support for introspection
I'm hoping that we can use this eventually to introspect config values.
The major driving idea here is to use this to get the command prefix
when you don't know it (`seabird: say command_prefix`).

This PR also adds support for querying startup time from Core as an
example of interacting with Core.